refactor(terminal): split TerminalMetadata.swift grab-bag into focused files
TerminalMetadata.swift (1806 LOC) held five unrelated concerns. The
struct keeps the file; frame meter, signposts, and both diagnostics
systems move to their own files. TerminalMetadataChangeClassifier moves
to AppState/Agent — it parses agent-status title heuristics and never
belonged in the terminal layer. All code moved verbatim.
