+++ Zentty/Terminal/TerminalMetadata.swift
@@ -1,315 +1,4 @@
 import Foundation
-import OSLog
-
-enum ZenttyPerformanceSignposts {
-    private static let signposter = OSSignposter(
-        subsystem: "be.zenjoy.zentty",
-        category: .pointsOfInterest
-    )
-
-    static func event(_ name: StaticString) {
-        signposter.emitEvent(name)
-    }
-
-    static func interval<T>(_ name: StaticString, _ body: () throws -> T) rethrows -> T {
-        let state = signposter.beginInterval(name)
-        defer {
-            signposter.endInterval(name, state)
-        }
-        return try body()
-    }
-}
-
-@MainActor
-final class TerminalFrameMeter {
-    enum SampleKind: String, Equatable {
-        case tick
-        case offset
-        case sent
-    }
-
-    enum Severity: Equatable {
-        case stable
-        case warning
-        case critical
-
-        var isDip: Bool {
-            self != .stable
-        }
-
-        static func classify(framesPerSecond: Double, preferredFramesPerSecond: Int) -> Severity {
-            let target = Double(max(1, preferredFramesPerSecond))
-            if framesPerSecond < target * 0.30 {
-                return .critical
-            }
-            if framesPerSecond < target * 0.50 {
-                return .warning
-            }
-            return .stable
-        }
-    }
-
-    struct HistoryPoint: Equatable {
-        let timestamp: TimeInterval
-        let framesPerSecond: Double?
-        let severity: Severity
-
-        var isDip: Bool {
-            severity.isDip
-        }
-    }
-
-    struct Sample: Equatable {
-        let paneID: PaneID
-        let timestamp: TimeInterval
-        let rowOffset: Double
-        let preferredFramesPerSecond: Int
-        let displayID: UInt32?
-        let isLiveScrolling: Bool
-        let sampleKind: SampleKind
-        let pacingMode: TerminalScrollFramePacingMode
-    }
-
-    struct Snapshot: Equatable {
-        let paneID: PaneID
-        let tickFramesPerSecond: Double?
-        let offsetFramesPerSecond: Double?
-        let sentFramesPerSecond: Double?
-        let preferredFramesPerSecond: Int
-        let lateFrameRatio: Double
-        let maxDeltaMilliseconds: Double?
-        let rowOffset: Double
-        let displayID: UInt32?
-        let isLiveScrolling: Bool
-        let sampleKind: SampleKind
-        let pacingMode: TerminalScrollFramePacingMode
-        let historyPoints: [HistoryPoint]
-
-        var framesPerSecond: Double? {
-            tickFramesPerSecond
-        }
-    }
-
-    static let shared = TerminalFrameMeter()
-    static let stateDidChangeNotification = Notification.Name("TerminalFrameMeterStateDidChange")
-
-    var isEnabled = false {
-        didSet {
-            guard oldValue != isEnabled else {
-                return
-            }
-            reset()
-            logger.notice("terminal frame meter \(self.isEnabled ? "enabled" : "disabled", privacy: .public)")
-            NotificationCenter.default.post(name: Self.stateDidChangeNotification, object: self)
-        }
-    }
-
-    private struct PaneStats {
-        var windowStartedAt: TimeInterval
-        var lastTimestamp: TimeInterval?
-        var tickFrameCount = 0
-        var offsetFrameCount = 0
-        var sentFrameCount = 0
-        var lateFrameCount = 0
-        var mainDeltaMaxMilliseconds: Double = 0
-        var latestRowOffset: Double = 0
-        var displayID: UInt32?
-        var preferredFramesPerSecond: Int
-        var isLiveScrolling = false
-        var sampleKind: SampleKind = .tick
-        var pacingMode: TerminalScrollFramePacingMode = .stopped
-        var lastTickTimestamp: TimeInterval?
-        var historyPoints: [HistoryPoint] = []
-    }
-
-    private let logger = Logger(subsystem: "be.zenjoy.zentty", category: "TerminalFrameMeter")
-    private var samples: [Sample] = []
-    private var statsByPane: [PaneID: PaneStats] = [:]
-    private var latestSnapshotsByPane: [PaneID: Snapshot] = [:]
-    private let sampleLimit = 2_000
-    private let historyWindowSeconds: TimeInterval = 5
-    private var nowOverride: TimeInterval?
-
-    @discardableResult
-    func recordScrollFrameSample(
-        paneID: PaneID,
-        rowOffset: Double,
-        preferredFramesPerSecond: Int,
-        displayID: UInt32?,
-        isLiveScrolling: Bool,
-        sampleKind: SampleKind,
-        pacingMode: TerminalScrollFramePacingMode
-    ) -> Snapshot? {
-        guard isEnabled else {
-            return nil
-        }
-
-        let now = currentTime()
-        let sample = Sample(
-            paneID: paneID,
-            timestamp: now,
-            rowOffset: rowOffset,
-            preferredFramesPerSecond: preferredFramesPerSecond,
-            displayID: displayID,
-            isLiveScrolling: isLiveScrolling,
-            sampleKind: sampleKind,
-            pacingMode: pacingMode
-        )
-        samples.append(sample)
-        if samples.count > sampleLimit {
-            samples.removeFirst(samples.count - sampleLimit)
-        }
-
-        return updateStats(with: sample)
-    }
-
-    private func updateStats(with sample: Sample) -> Snapshot {
-        let preferredFramesPerSecond = max(1, sample.preferredFramesPerSecond)
-        var stats = statsByPane[sample.paneID] ?? PaneStats(
-            windowStartedAt: sample.timestamp,
-            preferredFramesPerSecond: preferredFramesPerSecond
-        )
-
-        if sample.sampleKind == .tick {
-            if let lastTickTimestamp = stats.lastTickTimestamp {
-                let delta = sample.timestamp - lastTickTimestamp
-                let targetFrameDuration = 1.0 / Double(preferredFramesPerSecond)
-                if delta > targetFrameDuration * 1.5 {
-                    stats.lateFrameCount += 1
-                }
-                stats.mainDeltaMaxMilliseconds = max(stats.mainDeltaMaxMilliseconds, delta * 1_000)
-            }
-
-            let historyPoint = tickHistoryPoint(
-                timestamp: sample.timestamp,
-                previousTimestamp: stats.lastTickTimestamp,
-                preferredFramesPerSecond: preferredFramesPerSecond
-            )
-            stats.historyPoints.append(historyPoint)
-            stats.historyPoints.removeAll { sample.timestamp - $0.timestamp > historyWindowSeconds }
-            stats.lastTickTimestamp = sample.timestamp
-            stats.tickFrameCount += 1
-        }
-
-        switch sample.sampleKind {
-        case .tick:
-            break
-        case .offset:
-            stats.offsetFrameCount += 1
-        case .sent:
-            stats.sentFrameCount += 1
-        }
-        stats.lastTimestamp = sample.timestamp
-        stats.latestRowOffset = sample.rowOffset
-        stats.displayID = sample.displayID
-        stats.preferredFramesPerSecond = preferredFramesPerSecond
-        stats.isLiveScrolling = sample.isLiveScrolling
-        stats.sampleKind = sample.sampleKind
-        stats.pacingMode = sample.pacingMode
-
-        let elapsed = sample.timestamp - stats.windowStartedAt
-        let effectiveElapsed = elapsed > 0 ? elapsed : 1
-        let snapshot = Snapshot(
-            paneID: sample.paneID,
-            tickFramesPerSecond: framesPerSecond(count: stats.tickFrameCount, elapsed: effectiveElapsed),
-            offsetFramesPerSecond: framesPerSecond(count: stats.offsetFrameCount, elapsed: effectiveElapsed),
-            sentFramesPerSecond: framesPerSecond(count: stats.sentFrameCount, elapsed: effectiveElapsed),
-            preferredFramesPerSecond: preferredFramesPerSecond,
-            lateFrameRatio: stats.tickFrameCount > 0 ? Double(stats.lateFrameCount) / Double(stats.tickFrameCount) : 0,
-            maxDeltaMilliseconds: stats.mainDeltaMaxMilliseconds > 0 ? stats.mainDeltaMaxMilliseconds : nil,
-            rowOffset: stats.latestRowOffset,
-            displayID: stats.displayID,
-            isLiveScrolling: stats.isLiveScrolling,
-            sampleKind: stats.sampleKind,
-            pacingMode: stats.pacingMode,
-            historyPoints: stats.historyPoints
-        )
-        latestSnapshotsByPane[sample.paneID] = snapshot
-
-        if elapsed >= 1.0 {
-            #if DEBUG
-            let tickFPS = Double(stats.tickFrameCount) / elapsed
-            let offsetFPS = Double(stats.offsetFrameCount) / elapsed
-            let sentFPS = Double(stats.sentFrameCount) / elapsed
-            logger.log(
-                "pane=\(sample.paneID.rawValue, privacy: .public) tickFPS=\(tickFPS, format: .fixed(precision: 1)) offsetFPS=\(offsetFPS, format: .fixed(precision: 1)) sentFPS=\(sentFPS, format: .fixed(precision: 1)) target=\(preferredFramesPerSecond, privacy: .public) late=\(stats.lateFrameCount, privacy: .public)/\(stats.tickFrameCount, privacy: .public) maxDeltaMs=\(stats.mainDeltaMaxMilliseconds, format: .fixed(precision: 2)) rowOffset=\(stats.latestRowOffset, format: .fixed(precision: 2)) displayID=\(stats.displayID.map(String.init) ?? "nil", privacy: .public) live=\(stats.isLiveScrolling, privacy: .public) pacing=\(stats.pacingMode.rawValue, privacy: .public)"
-            )
-            #endif
-            stats = PaneStats(
-                windowStartedAt: sample.timestamp,
-                lastTimestamp: sample.timestamp,
-                preferredFramesPerSecond: preferredFramesPerSecond,
-                sampleKind: sample.sampleKind,
-                pacingMode: sample.pacingMode
-            )
-            stats.historyPoints = snapshot.historyPoints
-            stats.lastTickTimestamp = snapshot.historyPoints.last?.timestamp
-        }
-
-        statsByPane[sample.paneID] = stats
-        return snapshot
-    }
-
-    private func tickHistoryPoint(
-        timestamp: TimeInterval,
-        previousTimestamp: TimeInterval?,
-        preferredFramesPerSecond: Int
-    ) -> HistoryPoint {
-        guard let previousTimestamp else {
-            return HistoryPoint(timestamp: timestamp, framesPerSecond: nil, severity: .stable)
-        }
-
-        let delta = timestamp - previousTimestamp
-        guard delta > 0 else {
-            return HistoryPoint(timestamp: timestamp, framesPerSecond: nil, severity: .stable)
-        }
-
-        let framesPerSecond = 1.0 / delta
-        let severity = Severity.classify(
-            framesPerSecond: framesPerSecond,
-            preferredFramesPerSecond: preferredFramesPerSecond
-        )
-        return HistoryPoint(timestamp: timestamp, framesPerSecond: framesPerSecond, severity: severity)
-    }
-
-    private func framesPerSecond(count: Int, elapsed: TimeInterval) -> Double? {
-        guard count > 0 else {
-            return nil
-        }
-
-        return Double(count) / elapsed
-    }
-
-    private func reset() {
-        samples.removeAll(keepingCapacity: true)
-        statsByPane.removeAll(keepingCapacity: true)
-        latestSnapshotsByPane.removeAll(keepingCapacity: true)
-    }
-
-    private func currentTime() -> TimeInterval {
-        nowOverride ?? ProcessInfo.processInfo.systemUptime
-    }
-
-    func resetForTesting() {
-        reset()
-    }
-
-    func setNowForTesting(_ now: TimeInterval) {
-        nowOverride = now
-    }
-
-    func clearNowForTesting() {
-        nowOverride = nil
-    }
-
-    func samplesForTesting() -> [Sample] {
-        samples
-    }
-
-    func latestSnapshot(for paneID: PaneID) -> Snapshot? {
-        latestSnapshotsByPane[paneID]
-    }
-}
 
 struct TerminalMetadata: Equatable {
     var title: String?
@@ -329,1478 +18,4 @@ extension TerminalMetadata {
                 return !trimmed.isEmpty
             }
     }
-}
-
-enum TerminalMetadataChangeKind: Equatable {
-    case noop
-    case volatileTitleOnly
-    case meaningful
-}
-
-enum TerminalMetadataChangeClassifier {
-    enum VolatileAgentStatusPhase: Equatable {
-        case running
-        case starting
-        case needsInput
-        case idle
-    }
-
-    enum CodexWaitingTitleKind: Equatable {
-        case backgroundWait
-        case needsInput
-    }
-
-    struct VolatileAgentStatusTitleSignature: Equatable {
-        let phase: VolatileAgentStatusPhase
-        let subject: String
-    }
-
-    private struct ParsedVolatileAgentStatusTitle: Equatable {
-        let phase: VolatileAgentStatusPhase
-        let displaySubject: String
-    }
-
-    static func classify(previous: TerminalMetadata?, next: TerminalMetadata) -> TerminalMetadataChangeKind {
-        guard previous != next else {
-            return .noop
-        }
-
-        guard let previous else {
-            return .meaningful
-        }
-
-        guard previous.currentWorkingDirectory == next.currentWorkingDirectory,
-              previous.processName == next.processName,
-              previous.gitBranch == next.gitBranch else {
-            return .meaningful
-        }
-
-        let previousTool = AgentToolRecognizer.recognize(metadata: previous)
-        let nextTool = AgentToolRecognizer.recognize(metadata: next)
-        guard previousTool == nextTool else {
-            return .meaningful
-        }
-
-        if previousTool == .codex,
-           codexTaskProgress(for: previous.title) != codexTaskProgress(for: next.title) {
-            return .meaningful
-        }
-
-        if let previousSignature = realtimeAgentTitleSignature(
-            previous.title,
-            recognizedTool: previousTool
-        ),
-           let nextSignature = realtimeAgentTitleSignature(
-            next.title,
-            recognizedTool: nextTool
-           ),
-           previousSignature == nextSignature {
-            return .volatileTitleOnly
-        }
-
-        return .meaningful
-    }
-
-    static func volatileAgentStatusTitleSignature(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> VolatileAgentStatusTitleSignature? {
-        guard let rawNormalized = WorklaneContextFormatter.trimmed(value) else {
-            return nil
-        }
-
-        if recognizedTool == .hermes {
-            return parseHermesStatusTitle(rawNormalized)
-        }
-
-        guard recognizedTool == .codex else {
-            return nil
-        }
-
-        let normalized = stripTrailingCodexTaskProgress(from: rawNormalized)
-        if let threadStatus = parseCodexThreadStatusTitle(normalized) {
-            return VolatileAgentStatusTitleSignature(
-                phase: .needsInput,
-                subject: threadStatus.subject
-            )
-        }
-        if let actionRequired = parseCodexActionRequiredTitle(normalized) {
-            return VolatileAgentStatusTitleSignature(
-                phase: .needsInput,
-                subject: actionRequired.subject
-            )
-        }
-
-        guard let parsed = parseAgentStatusTitle(
-                  normalized,
-                  runningWords: ["working", "thinking"],
-                  startingWords: ["starting"],
-                  needsInputWords: ["waiting"],
-                  idleWords: ["ready"]
-              ) else {
-            return nil
-        }
-
-        let phase: VolatileAgentStatusPhase
-        switch codexWaitingTitleKind(for: rawNormalized) {
-        case .backgroundWait:
-            phase = .idle
-        case .needsInput:
-            phase = .needsInput
-        case nil:
-            phase = parsed.phase
-        }
-
-        return VolatileAgentStatusTitleSignature(
-            phase: phase,
-            subject: parsed.displaySubject.lowercased()
-        )
-    }
-
-    static func volatileAgentStatusDisplaySubject(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> String? {
-        guard recognizedTool == .codex,
-              let rawNormalized = WorklaneContextFormatter.trimmed(value) else {
-            return nil
-        }
-        let normalized = stripTrailingCodexTaskProgress(from: rawNormalized)
-        guard let parsed = parseAgentStatusTitle(
-                  normalized,
-                  runningWords: ["working", "thinking"],
-                  startingWords: ["starting"],
-                  needsInputWords: ["waiting"],
-                  idleWords: ["ready"]
-              ) else {
-            return nil
-        }
-
-        guard parsed.displaySubject.caseInsensitiveCompare("zentty") != .orderedSame else {
-            return nil
-        }
-
-        return parsed.displaySubject
-    }
-
-    static func diagnosticAgentStatusTitleSignature(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> VolatileAgentStatusTitleSignature? {
-        guard let normalized = WorklaneContextFormatter.trimmed(value) else {
-            return nil
-        }
-
-        switch recognizedTool {
-        case .codex:
-            return volatileAgentStatusTitleSignature(normalized, recognizedTool: .codex)
-        case .claudeCode:
-            // Claude Code 2.x encodes status in the title glyph: "✳" (U+2733)
-            // on the idle prompt, braille spinner glyphs (U+2800…U+28FF) while
-            // the agent is thinking. After a user interrupt (Escape) the
-            // spinner is replaced by "✳", with the subject left intact — this
-            // is what lets us detect the interrupt when no Stop hook fires.
-            if let signature = parseClaudeCodeGlyphTitle(normalized) {
-                return signature
-            }
-
-            // Fallback: older Claude Code builds used English words
-            // ("Thinking …", "Interrupted · …") as the title prefix. Keep
-            // detection for those so downgrades stay covered.
-            guard let parsed = parseAgentStatusTitle(
-                normalized,
-                runningWords: ["thinking", "working", "responding", "analyzing"],
-                startingWords: ["starting"],
-                idleWords: ["ready", "waiting", "interrupted"]
-            ) else {
-                return nil
-            }
-
-            return VolatileAgentStatusTitleSignature(
-                phase: parsed.phase,
-                subject: parsed.displaySubject.lowercased()
-            )
-        default:
-            return nil
-        }
-    }
-
-    static func realtimeAgentTitleSignature(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> VolatileAgentStatusTitleSignature? {
-        switch recognizedTool {
-        case .codex:
-            return volatileAgentStatusTitleSignature(value, recognizedTool: .codex)
-        case .hermes:
-            return volatileAgentStatusTitleSignature(value, recognizedTool: .hermes)
-        case .claudeCode:
-            return diagnosticAgentStatusTitleSignature(value, recognizedTool: .claudeCode)
-        default:
-            return nil
-        }
-    }
-
-    static func wouldTreatAsVolatileClaudeTransition(
-        previous: TerminalMetadata?,
-        next: TerminalMetadata
-    ) -> Bool {
-        guard let previous else {
-            return false
-        }
-
-        let previousTool = AgentToolRecognizer.recognize(metadata: previous)
-        let nextTool = AgentToolRecognizer.recognize(metadata: next)
-        guard previousTool == .claudeCode, nextTool == .claudeCode else {
-            return false
-        }
-
-        guard previous.title != next.title else {
-            return false
-        }
-
-        return diagnosticAgentStatusTitleSignature(previous.title, recognizedTool: previousTool)
-            == diagnosticAgentStatusTitleSignature(next.title, recognizedTool: nextTool)
-    }
-
-    static func isVolatileAgentStatusTitle(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> Bool {
-        volatileAgentStatusTitleSignature(value, recognizedTool: recognizedTool) != nil
-    }
-
-    static func isRealtimeAgentStatusTitle(
-        _ value: String?,
-        recognizedTool: AgentTool?
-    ) -> Bool {
-        realtimeAgentTitleSignature(value, recognizedTool: recognizedTool) != nil
-    }
-
-    static func codexWaitingTitleKind(for value: String?) -> CodexWaitingTitleKind? {
-        guard let rawNormalized = WorklaneContextFormatter.trimmed(value) else {
-            return nil
-        }
-        let normalized = stripTrailingCodexTaskProgress(from: rawNormalized)
-
-        let firstWord = normalized.prefix(while: { $0.isLetter }).lowercased()
-        guard firstWord == "waiting" else {
-            return nil
-        }
-
-        if AgentInteractionClassifier.requiresHumanInput(message: normalized) {
-            return .needsInput
-        }
-
-        return .backgroundWait
-    }
-
-    static func codexTitleInteractionKind(for value: String?) -> PaneAgentInteractionKind? {
-        guard let rawNormalized = WorklaneContextFormatter.trimmed(value) else {
-            return nil
-        }
-        let normalized = stripTrailingCodexTaskProgress(from: rawNormalized)
-
-        if let threadStatus = parseCodexThreadStatusTitle(normalized) {
-            return threadStatus.interactionKind
-        }
-        if let actionRequired = parseCodexActionRequiredTitle(normalized) {
-            return actionRequired.interactionKind
-        }
-
-        guard codexWaitingTitleKind(for: rawNormalized) == .needsInput else {
-            return nil
-        }
-
-        return AgentInteractionClassifier.interactionKind(forWaitingMessage: normalized)
-            ?? .genericInput
-    }
-
-    static func codexTaskProgress(for value: String?) -> PaneAgentTaskProgress? {
-        guard let normalized = WorklaneContextFormatter.trimmed(value),
-              let (_, progress) = splitTrailingCodexTaskProgress(from: normalized) else {
-            return nil
-        }
-
-        return progress
-    }
-
-    private static func parseClaudeCodeGlyphTitle(
-        _ normalized: String
-    ) -> VolatileAgentStatusTitleSignature? {
-        guard let first = normalized.unicodeScalars.first else {
-            return nil
-        }
-
-        let phase: VolatileAgentStatusPhase
-        switch first.value {
-        case 0x2733:
-            phase = .idle
-        case 0x2800...0x28FF:
-            phase = .running
-        default:
-            return nil
-        }
-
-        let remainder = normalized
-            .dropFirst()
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !remainder.isEmpty else {
-            return nil
-        }
-
-        return VolatileAgentStatusTitleSignature(
-            phase: phase,
-            subject: remainder.lowercased()
-        )
-    }
-
-    private static func parseCodexThreadStatusTitle(
-        _ normalized: String
-    ) -> (subject: String, interactionKind: PaneAgentInteractionKind)? {
-        let words = normalized
-            .split(whereSeparator: { !$0.isLetter })
-            .prefix(3)
-            .map { String($0).lowercased() }
-        guard words.count == 3,
-              ["main", "parent"].contains(words[0]),
-              words[1] == "needs" else {
-            return nil
-        }
-
-        let interactionKind: PaneAgentInteractionKind
-        switch words[2] {
-        case "approval":
-            return nil
-        case "input":
-            interactionKind = .genericInput
-        default:
-            return nil
-        }
-
-        return (normalized.lowercased(), interactionKind)
-    }
-
-    private static func parseCodexActionRequiredTitle(
-        _ normalized: String
-    ) -> (subject: String, interactionKind: PaneAgentInteractionKind)? {
-        let stripped = stripCodexTitleBadge(from: normalized)
-        let actionRequiredPrefix = "action required"
-        guard stripped.lowercased().hasPrefix(actionRequiredPrefix) else {
-            return nil
-        }
-
-        return (stripped.lowercased(), .genericInput)
-    }
-
-    private static func stripCodexTitleBadge(from value: String) -> String {
-        var remainder = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard remainder.hasPrefix("["),
-              let closingIndex = remainder.firstIndex(of: "]") else {
-            return remainder
-        }
-        let badge = remainder[remainder.index(after: remainder.startIndex)..<closingIndex]
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard badge == "!" || badge == "." else {
-            return remainder
-        }
-        remainder = String(remainder[remainder.index(after: closingIndex)...])
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        return remainder
-    }
-
-    private static func parseAgentStatusTitle(
-        _ normalized: String,
-        runningWords: Set<String>,
-        startingWords: Set<String>,
-        needsInputWords: Set<String> = [],
-        idleWords: Set<String>
-    ) -> ParsedVolatileAgentStatusTitle? {
-        let firstWord = normalized.prefix(while: { $0.isLetter }).lowercased()
-        let phase: VolatileAgentStatusPhase
-        if runningWords.contains(firstWord) {
-            phase = .running
-        } else if startingWords.contains(firstWord) {
-            phase = .starting
-        } else if needsInputWords.contains(firstWord) {
-            phase = .needsInput
-        } else if idleWords.contains(firstWord) {
-            phase = .idle
-        } else {
-            return nil
-        }
-
-        var remainder = normalized.dropFirst(firstWord.count)
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !remainder.isEmpty else {
-            return nil
-        }
-
-        if let firstToken = remainder.split(whereSeparator: \.isWhitespace).first,
-           firstToken.contains(where: \.isLetter) == false,
-           firstToken.contains(where: \.isNumber) == false {
-            remainder = remainder.dropFirst(firstToken.count)
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-        }
-
-        guard !remainder.isEmpty else {
-            return nil
-        }
-
-        return ParsedVolatileAgentStatusTitle(
-            phase: phase,
-            displaySubject: remainder
-        )
-    }
-
-    private static func parseHermesStatusTitle(
-        _ normalized: String
-    ) -> VolatileAgentStatusTitleSignature? {
-        guard let firstCharacter = normalized.first,
-              let firstScalar = firstCharacter.unicodeScalars.first else {
-            return nil
-        }
-
-        let phase: VolatileAgentStatusPhase
-        switch firstScalar.value {
-        case 0x23F3: // ⏳
-            phase = .running
-        case 0x2713: // ✓
-            phase = .idle
-        case 0x26A0: // ⚠ / ⚠️
-            phase = .needsInput
-        default:
-            return nil
-        }
-
-        let subject = normalized.dropFirst()
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !subject.isEmpty else {
-            return nil
-        }
-
-        return VolatileAgentStatusTitleSignature(
-            phase: phase,
-            subject: subject.lowercased()
-        )
-    }
-
-    private static func stripTrailingCodexTaskProgress(from value: String) -> String {
-        splitTrailingCodexTaskProgress(from: value)?.title ?? value
-    }
-
-    private static func splitTrailingCodexTaskProgress(
-        from value: String
-    ) -> (title: String, progress: PaneAgentTaskProgress)? {
-        guard let tasksRange = value.range(of: " | Tasks ", options: .backwards) else {
-            return nil
-        }
-
-        let title = value[..<tasksRange.lowerBound].trimmingCharacters(in: .whitespacesAndNewlines)
-        let progressValue = value[tasksRange.upperBound...].trimmingCharacters(in: .whitespacesAndNewlines)
-        let parts = progressValue.split(separator: "/", omittingEmptySubsequences: false)
-        guard parts.count == 2,
-              let doneCount = Int(parts[0]),
-              let totalCount = Int(parts[1]),
-              let progress = PaneAgentTaskProgress(doneCount: doneCount, totalCount: totalCount),
-              !title.isEmpty else {
-            return nil
-        }
-
-        return (title: title, progress: progress)
-    }
-}
-
-enum TerminalMetadataDeliveryOutcome: String, Equatable, Sendable {
-    case immediate
-    case coalesced
-    case droppedNoop = "dropped_noop"
-}
-
-enum TerminalDiagnosticsRenderKind: String, Sendable {
-    case full
-    case sidebar
-    case header
-    case canvas
-}
-
-enum TerminalViewportLaneRole: String, Equatable, Sendable {
-    case activeCanvas
-    case peekNeighbor
-    case unknown
-}
-
-enum TerminalViewportEventSource: String, Equatable, Sendable {
-    case activeZoomIn
-    case activeZoomInUnsuspend
-    case forceViewportSync
-    case initialSuspension
-    case libghosttySurfaceInit
-    case libghosttyUpdateViewport
-    case paneContainerFrameSync
-    case paneContainerReparent
-    case paneContainerSessionActivate
-    case paneMounted
-    case peekCommit
-    case peekNeighborBind
-    case peekNeighborCenterOnPane
-    case peekNeighborAbandonZoomOut
-    case peekNeighborDetach
-    case peekNeighborEndZoomOut
-    case peekNeighborPrepareZoomOut
-    case peekNeighborResetFullCanvas
-    case peekOpened
-    case peekViewDetach
-    case rootPeekDidClose
-    case scrollHostLayout
-    case scrollHostSync
-    case scrollHostSyncSuspended
-    case syncAttempt
-    case syncSkippedDuplicate
-    case syncSkippedNoWindow
-    case syncSkippedSuspended
-    case syncSkippedZeroBounds
-    case syncSuspended
-    case syncUnsuspended
-}
-
-@MainActor
-protocol TerminalViewportDiagnosticsContextConfiguring: AnyObject {
-    func updateViewportDiagnosticsContext(_ context: TerminalViewportDiagnostics.Context)
-}
-
-final class TerminalViewportDiagnostics: @unchecked Sendable {
-    struct Context: Equatable, Sendable {
-        var paneID: PaneID?
-        var worklaneID: WorklaneID?
-        var laneRole: TerminalViewportLaneRole
-        var isZoomedOut: Bool?
-        var isViewportSyncSuspended: Bool?
-        var windowAttached: Bool?
-        var containerBounds: CGRect?
-        var terminalHostBounds: CGRect?
-        var scrollHostBounds: CGRect?
-        var surfaceBounds: CGRect?
-        var viewportSize: CGSize?
-        var previousViewportSize: CGSize?
-        var scale: CGFloat?
-        var displayID: UInt32?
-        var peekSessionID: String?
-        var note: String?
-
-        init(
-            paneID: PaneID? = nil,
-            worklaneID: WorklaneID? = nil,
-            laneRole: TerminalViewportLaneRole = .unknown,
-            isZoomedOut: Bool? = nil,
-            isViewportSyncSuspended: Bool? = nil,
-            windowAttached: Bool? = nil,
-            containerBounds: CGRect? = nil,
-            terminalHostBounds: CGRect? = nil,
-            scrollHostBounds: CGRect? = nil,
-            surfaceBounds: CGRect? = nil,
-            viewportSize: CGSize? = nil,
-            previousViewportSize: CGSize? = nil,
-            scale: CGFloat? = nil,
-            displayID: UInt32? = nil,
-            peekSessionID: String? = nil,
-            note: String? = nil
-        ) {
-            self.paneID = paneID
-            self.worklaneID = worklaneID
-            self.laneRole = laneRole
-            self.isZoomedOut = isZoomedOut
-            self.isViewportSyncSuspended = isViewportSyncSuspended
-            self.windowAttached = windowAttached
-            self.containerBounds = containerBounds
-            self.terminalHostBounds = terminalHostBounds
-            self.scrollHostBounds = scrollHostBounds
-            self.surfaceBounds = surfaceBounds
-            self.viewportSize = viewportSize
-            self.previousViewportSize = previousViewportSize
-            self.scale = scale
-            self.displayID = displayID
-            self.peekSessionID = peekSessionID
-            self.note = note
-        }
-    }
-
-    struct Event: Equatable, Sendable {
-        let sequence: Int
-        let timestamp: Date
-        let source: TerminalViewportEventSource
-        let context: Context
-    }
-
-    static let shared: TerminalViewportDiagnostics = {
-        let diagnostics = TerminalViewportDiagnostics()
-        diagnostics.setEnabled(defaultSharedEnabled)
-        return diagnostics
-    }()
-
-    var onRecord: ((Event) -> Void)?
-
-    private static let logger = Logger(subsystem: "be.zenjoy.zentty", category: "WorklanePeekViewport")
-    private let lock = NSLock()
-    private let maxEvents: Int
-    private var enabled = false
-    private var nextSequence = 1
-    private var events: [Event] = []
-
-    init(maxEvents: Int = 2_000) {
-        self.maxEvents = max(1, maxEvents)
-    }
-
-    func setEnabled(_ enabled: Bool) {
-        lock.lock()
-        self.enabled = enabled
-        lock.unlock()
-    }
-
-    func record(_ source: TerminalViewportEventSource, context: Context = Context()) {
-        lock.lock()
-        guard enabled else {
-            lock.unlock()
-            return
-        }
-
-        let event = Event(
-            sequence: nextSequence,
-            timestamp: Date(),
-            source: source,
-            context: context
-        )
-        nextSequence += 1
-        events.append(event)
-        if events.count > maxEvents {
-            events.removeFirst(events.count - maxEvents)
-        }
-        let onRecord = self.onRecord
-        lock.unlock()
-
-        Self.logger.log("\(Self.logPayload(for: event), privacy: .public)")
-        onRecord?(event)
-    }
-
-    #if DEBUG
-    func clearForTesting() {
-        lock.lock()
-        events.removeAll()
-        nextSequence = 1
-        lock.unlock()
-    }
-
-    func eventsForTesting() -> [Event] {
-        lock.lock()
-        defer { lock.unlock() }
-        return events
-    }
-
-    static func logPayloadForTesting(_ event: Event) -> String {
-        logPayload(for: event)
-    }
-    #endif
-
-    private static var defaultSharedEnabled: Bool {
-        #if DEBUG
-        ProcessInfo.processInfo.environment["ZENTTY_WORKLANE_PEEK_DIAGNOSTICS"] == "1"
-        #else
-        false
-        #endif
-    }
-
-    private static func logPayload(for event: Event) -> String {
-        let context = event.context
-        var components: [String] = [
-            "seq=\(event.sequence)",
-            "source=\(event.source.rawValue)",
-            "laneRole=\(context.laneRole.rawValue)",
-        ]
-
-        if let paneID = context.paneID {
-            components.append("paneID=\(paneID.rawValue)")
-        }
-        if let worklaneID = context.worklaneID {
-            components.append("worklaneID=\(worklaneID.rawValue)")
-        }
-        if let isZoomedOut = context.isZoomedOut {
-            components.append("isZoomedOut=\(isZoomedOut)")
-        }
-        if let isViewportSyncSuspended = context.isViewportSyncSuspended {
-            components.append("isViewportSyncSuspended=\(isViewportSyncSuspended)")
-        }
-        if let windowAttached = context.windowAttached {
-            components.append("windowAttached=\(windowAttached)")
-        }
-        if let containerBounds = context.containerBounds {
-            components.append("container=\(format(rect: containerBounds))")
-        }
-        if let terminalHostBounds = context.terminalHostBounds {
-            components.append("terminalHost=\(format(rect: terminalHostBounds))")
-        }
-        if let scrollHostBounds = context.scrollHostBounds {
-            components.append("scrollHost=\(format(rect: scrollHostBounds))")
-        }
-        if let surfaceBounds = context.surfaceBounds {
-            components.append("surface=\(format(rect: surfaceBounds))")
-        }
-        if let previousViewportSize = context.previousViewportSize {
-            components.append("previousViewport=\(format(size: previousViewportSize))")
-        }
-        if let viewportSize = context.viewportSize {
-            components.append("viewport=\(format(size: viewportSize))")
-        }
-        if let scale = context.scale {
-            components.append("scale=\(scale)")
-        }
-        if let displayID = context.displayID {
-            components.append("displayID=\(displayID)")
-        }
-        if let peekSessionID = context.peekSessionID {
-            components.append("peekSessionID=\(peekSessionID)")
-        }
-        if let note = context.note {
-            components.append("note=\(note)")
-        }
-
-        return components.joined(separator: " ")
-    }
-
-    private static func format(size: CGSize) -> String {
-        "\(size.width)x\(size.height)"
-    }
-
-    private static func format(rect: CGRect) -> String {
-        "\(rect.origin.x),\(rect.origin.y),\(rect.width)x\(rect.height)"
-    }
-}
-
-private let terminalDiagnosticsLogger = Logger(subsystem: "be.zenjoy.zentty", category: "TerminalDiagnostics")
-
-final class TerminalDiagnostics: @unchecked Sendable {
-    enum Scope: Hashable, Equatable, Sendable {
-        case runtime
-        case pane(PaneID)
-
-        var label: String {
-            switch self {
-            case .runtime:
-                return "runtime"
-            case .pane(let paneID):
-                return "pane:\(paneID.rawValue)"
-            }
-        }
-    }
-
-    struct BurstSummary: Equatable, Sendable {
-        let scope: Scope
-        var startedAt: Date
-        var endedAt: Date
-        var commandHint: String?
-        var wakeupCount = 0
-        var wakeupEnqueueCount = 0
-        var tickCount = 0
-        var tickTotalMilliseconds: Double = 0
-        var tickMaxMilliseconds: Double = 0
-        var mainQueueDelayTotalMilliseconds: Double = 0
-        var mainQueueDelayMaxMilliseconds: Double = 0
-        var actionCallbackCount = 0
-        var actionDrainCount = 0
-        var actionDrainQueueDelayTotalMilliseconds: Double = 0
-        var actionDrainQueueDelayMaxMilliseconds: Double = 0
-        var scrollbarApplyCount = 0
-        var scrollbarApplyTotalMilliseconds: Double = 0
-        var scrollbarApplyMaxMilliseconds: Double = 0
-        var scrollHostSyncCount = 0
-        var scrollHostSyncTotalMilliseconds: Double = 0
-        var scrollHostSyncMaxMilliseconds: Double = 0
-        var scrollbarGeometryApplyCount = 0
-        var documentHeightChangeCount = 0
-        var documentHeightMaxPoints: Double = 0
-        var documentHeightMaxDeltaPoints: Double = 0
-        var scrollbarBottomAlignedCount = 0
-        var scrollbarOffBottomCount = 0
-        var scrollbarWasAtBottomCount = 0
-        var scrollbarAutoFollowEligibleCount = 0
-        var scrollbarAutoFollowSuppressedCount = 0
-        var scrollbarAutoFollowAppliedCount = 0
-        var scrollbarUserScrolledAwayCount = 0
-        var scrollbarExplicitSyncAllowedCount = 0
-        var scrollbarMaxTotalRows: UInt64 = 0
-        var scrollbarMinOffsetRows: Double?
-        var scrollbarMaxOffsetRows: Double = 0
-        var scrollbarMinVisibleRows: UInt64?
-        var scrollbarMaxVisibleRows: UInt64 = 0
-        var firstScrollbarPosition: String?
-        var lastScrollbarPosition: String?
-        var replaySuppressionEnterCount = 0
-        var replaySuppressionExitCount = 0
-        var replaySuppressionTotalMilliseconds: Double = 0
-        var replaySuppressionReasonCounts: [String: Int] = [:]
-        var occlusionVisibleCount = 0
-        var occlusionHiddenCount = 0
-        var reflectScrolledClipViewCount = 0
-        var scrollToRowActionCount = 0
-        var viewportSyncCount = 0
-        var viewportSyncTotalMilliseconds: Double = 0
-        var viewportSyncMaxMilliseconds: Double = 0
-        var actionPayloadCounts: [String: Int] = [:]
-        var metadataDeliveryCount = 0
-        var metadataChangeKindCounts: [String: Int] = [:]
-        var metadataToolCounts: [String: Int] = [:]
-        var titleChangeCount = 0
-        var claudeTitleChangeCount = 0
-        var codexTitleChangeCount = 0
-        var wouldHaveBeenVolatileClaudeCount = 0
-        var storeUpdateCount = 0
-        var storeFastPathCount = 0
-        var auxiliaryInvalidationCounts: [String: Int] = [:]
-        var renderCounts: [String: Int] = [:]
-        var visiblePaneCount: Int?
-        var liveRuntimeCount: Int?
-        var windowVisible: Bool?
-        var windowKey: Bool?
-        var firstTitle: String?
-        var lastTitle: String?
-        var uniqueTitleCount = 0
-
-        fileprivate var uniqueTitles: Set<String> = []
-
-        fileprivate mutating func recordTitle(_ title: String?) {
-            guard let title = WorklaneContextFormatter.trimmed(title) else {
-                return
-            }
-
-            if firstTitle == nil {
-                firstTitle = title
-            }
-            lastTitle = title
-            if uniqueTitles.insert(title).inserted {
-                uniqueTitleCount = uniqueTitles.count
-            }
-        }
-
-    }
-
-    static let shared: TerminalDiagnostics = {
-        let diagnostics = TerminalDiagnostics()
-        diagnostics.setEnabled(defaultSharedEnabled)
-        return diagnostics
-    }()
-
-    var onEmit: ((BurstSummary) -> Void)?
-
-    private struct BurstState {
-        var summary: BurstSummary
-        var lastEventAt: Date
-        var flushWorkItem: DispatchWorkItem?
-    }
-
-    private let quietPeriod: TimeInterval
-    private let queue = DispatchQueue(label: "be.zenjoy.zentty.terminal-diagnostics", qos: .utility)
-    private let deliverEmission: (@escaping () -> Void) -> Void
-    private let enabledLock = NSLock()
-    private var enabled = false
-    private var burstsByScope: [Scope: BurstState] = [:]
-
-    init(
-        quietPeriod: TimeInterval = 1.0,
-        deliverEmission: @escaping (@escaping () -> Void) -> Void = {
-            DispatchQueue.main.async(execute: $0)
-        }
-    ) {
-        self.quietPeriod = quietPeriod
-        self.deliverEmission = deliverEmission
-    }
-
-    func setEnabled(_ enabled: Bool) {
-        enabledLock.lock()
-        self.enabled = enabled
-        enabledLock.unlock()
-
-        guard enabled == false else {
-            return
-        }
-
-        queue.async { [weak self] in
-            guard let self else {
-                return
-            }
-            guard self.isEnabled == false else {
-                return
-            }
-
-            self.burstsByScope.values.forEach { $0.flushWorkItem?.cancel() }
-            self.burstsByScope.removeAll()
-        }
-    }
-
-    func recordWakeupReceived() {
-        record(scope: .runtime) { summary in
-            summary.wakeupCount += 1
-        }
-    }
-
-    func recordWakeupEnqueued() {
-        record(scope: .runtime) { summary in
-            summary.wakeupEnqueueCount += 1
-        }
-    }
-
-    func recordTick(durationNanoseconds: UInt64, queueDelayNanoseconds: UInt64) {
-        record(scope: .runtime) { summary in
-            summary.tickCount += 1
-            let duration = Self.milliseconds(from: durationNanoseconds)
-            summary.tickTotalMilliseconds += duration
-            summary.tickMaxMilliseconds = max(summary.tickMaxMilliseconds, duration)
-            let queueDelay = Self.milliseconds(from: queueDelayNanoseconds)
-            summary.mainQueueDelayTotalMilliseconds += queueDelay
-            summary.mainQueueDelayMaxMilliseconds = max(summary.mainQueueDelayMaxMilliseconds, queueDelay)
-        }
-    }
-
-    func recordActionCallback(paneID: PaneID, payload: LibghosttySurfaceActionPayload) {
-        record(scope: .pane(paneID)) { summary in
-            summary.actionCallbackCount += 1
-            let key = Self.actionPayloadKey(for: payload)
-            Self.increment(key, in: &summary.actionPayloadCounts)
-            if case .setTitle(let title) = payload {
-                summary.recordTitle(title)
-            }
-        }
-    }
-
-    func recordActionDrain(paneID: PaneID, queueDelayNanoseconds: UInt64 = 0) {
-        record(scope: .pane(paneID)) { summary in
-            summary.actionDrainCount += 1
-            let queueDelay = Self.milliseconds(from: queueDelayNanoseconds)
-            summary.actionDrainQueueDelayTotalMilliseconds += queueDelay
-            summary.actionDrainQueueDelayMaxMilliseconds = max(
-                summary.actionDrainQueueDelayMaxMilliseconds,
-                queueDelay
-            )
-        }
-    }
-
-    func recordScrollbarApply(paneID: PaneID, durationNanoseconds: UInt64) {
-        record(scope: .pane(paneID)) { summary in
-            summary.scrollbarApplyCount += 1
-            let duration = Self.milliseconds(from: durationNanoseconds)
-            summary.scrollbarApplyTotalMilliseconds += duration
-            summary.scrollbarApplyMaxMilliseconds = max(summary.scrollbarApplyMaxMilliseconds, duration)
-        }
-    }
-
-    func recordScrollHostSync(
-        paneID: PaneID,
-        durationNanoseconds: UInt64,
-        geometryApplied: Bool,
-        documentHeightChanged: Bool,
-        documentHeightPoints: CGFloat,
-        documentHeightDeltaPoints: CGFloat,
-        reflected: Bool,
-        scrollbarTotalRows: UInt64?,
-        scrollbarOffsetRows: Double?,
-        scrollbarVisibleRows: UInt64?,
-        wasAtBottom: Bool?,
-        shouldAutoScroll: Bool?,
-        autoScrollApplied: Bool?,
-        userScrolledAwayFromBottom: Bool?,
-        explicitScrollbarSyncAllowed: Bool?
-    ) {
-        record(scope: .pane(paneID)) { summary in
-            summary.scrollHostSyncCount += 1
-            let duration = Self.milliseconds(from: durationNanoseconds)
-            summary.scrollHostSyncTotalMilliseconds += duration
-            summary.scrollHostSyncMaxMilliseconds = max(summary.scrollHostSyncMaxMilliseconds, duration)
-            if geometryApplied {
-                summary.scrollbarGeometryApplyCount += 1
-            }
-            if documentHeightChanged {
-                summary.documentHeightChangeCount += 1
-            }
-            summary.documentHeightMaxPoints = max(summary.documentHeightMaxPoints, Double(documentHeightPoints))
-            summary.documentHeightMaxDeltaPoints = max(summary.documentHeightMaxDeltaPoints, Double(documentHeightDeltaPoints))
-            if reflected {
-                summary.reflectScrolledClipViewCount += 1
-            }
-            if let total = scrollbarTotalRows, let offset = scrollbarOffsetRows, let visible = scrollbarVisibleRows {
-                let position = "total:\(total),offset:\(offset),len:\(visible)"
-                if summary.firstScrollbarPosition == nil {
-                    summary.firstScrollbarPosition = position
-                }
-                summary.lastScrollbarPosition = position
-                summary.scrollbarMaxTotalRows = max(summary.scrollbarMaxTotalRows, total)
-                summary.scrollbarMinOffsetRows = min(summary.scrollbarMinOffsetRows ?? offset, offset)
-                summary.scrollbarMaxOffsetRows = max(summary.scrollbarMaxOffsetRows, offset)
-                summary.scrollbarMinVisibleRows = min(summary.scrollbarMinVisibleRows ?? visible, visible)
-                summary.scrollbarMaxVisibleRows = max(summary.scrollbarMaxVisibleRows, visible)
-                if Self.rowsBelowViewport(total: total, offset: offset, visible: visible) == 0 {
-                    summary.scrollbarBottomAlignedCount += 1
-                } else {
-                    summary.scrollbarOffBottomCount += 1
-                }
-            }
-            if wasAtBottom == true {
-                summary.scrollbarWasAtBottomCount += 1
-            }
-            if shouldAutoScroll == true {
-                summary.scrollbarAutoFollowEligibleCount += 1
-            } else if shouldAutoScroll == false {
-                summary.scrollbarAutoFollowSuppressedCount += 1
-            }
-            if autoScrollApplied == true {
-                summary.scrollbarAutoFollowAppliedCount += 1
-            }
-            if userScrolledAwayFromBottom == true {
-                summary.scrollbarUserScrolledAwayCount += 1
-            }
-            if explicitScrollbarSyncAllowed == true {
-                summary.scrollbarExplicitSyncAllowedCount += 1
-            }
-        }
-    }
-
-    func recordScrollToRowAction(paneID: PaneID) {
-        record(scope: .pane(paneID)) { summary in
-            summary.scrollToRowActionCount += 1
-        }
-    }
-
-    func recordReplaySuppressionEntered(paneID: PaneID, reason: String) {
-        record(scope: .pane(paneID)) { summary in
-            summary.replaySuppressionEnterCount += 1
-            Self.increment(reason, in: &summary.replaySuppressionReasonCounts)
-        }
-    }
-
-    func recordReplaySuppressionExited(paneID: PaneID, reason: String, durationNanoseconds: UInt64) {
-        record(scope: .pane(paneID)) { summary in
-            summary.replaySuppressionExitCount += 1
-            summary.replaySuppressionTotalMilliseconds += Self.milliseconds(from: durationNanoseconds)
-            Self.increment(reason, in: &summary.replaySuppressionReasonCounts)
-        }
-    }
-
-    func recordOcclusionVisibility(paneID: PaneID, visible: Bool) {
-        record(scope: .pane(paneID)) { summary in
-            if visible {
-                summary.occlusionVisibleCount += 1
-            } else {
-                summary.occlusionHiddenCount += 1
-            }
-        }
-    }
-
-    func recordViewportSync(paneID: PaneID, durationNanoseconds: UInt64) {
-        record(scope: .pane(paneID)) { summary in
-            summary.viewportSyncCount += 1
-            let duration = Self.milliseconds(from: durationNanoseconds)
-            summary.viewportSyncTotalMilliseconds += duration
-            summary.viewportSyncMaxMilliseconds = max(summary.viewportSyncMaxMilliseconds, duration)
-        }
-    }
-
-    func recordMetadataObservation(
-        paneID: PaneID,
-        previous: TerminalMetadata?,
-        next: TerminalMetadata,
-        changeKind: TerminalMetadataChangeKind,
-        delivery: TerminalMetadataDeliveryOutcome
-    ) {
-        record(scope: .pane(paneID)) { summary in
-            Self.increment(Self.metadataChangeKindKey(changeKind), in: &summary.metadataChangeKindCounts)
-            let tool = AgentToolRecognizer.recognize(metadata: next)
-            Self.increment(Self.toolKey(tool), in: &summary.metadataToolCounts)
-            if summary.commandHint == nil {
-                if let tool {
-                    summary.commandHint = Self.toolKey(tool)
-                } else {
-                    summary.commandHint = WorklaneContextFormatter.trimmed(next.processName)
-                }
-            }
-
-            if previous?.title != next.title {
-                summary.titleChangeCount += 1
-                switch tool {
-                case .claudeCode:
-                    summary.claudeTitleChangeCount += 1
-                case .codex:
-                    summary.codexTitleChangeCount += 1
-                default:
-                    break
-                }
-            }
-
-            if TerminalMetadataChangeClassifier.wouldTreatAsVolatileClaudeTransition(
-                previous: previous,
-                next: next
-            ) {
-                summary.wouldHaveBeenVolatileClaudeCount += 1
-            }
-
-            summary.recordTitle(next.title)
-        }
-    }
-
-    func recordMetadataDelivery(paneID: PaneID, outcome _: TerminalMetadataDeliveryOutcome) {
-        record(scope: .pane(paneID)) { summary in
-            summary.metadataDeliveryCount += 1
-        }
-    }
-
-    func recordStoreMetadataUpdate(paneID: PaneID) {
-        record(scope: .pane(paneID)) { summary in
-            summary.storeUpdateCount += 1
-        }
-    }
-
-    func recordStoreFastPath(paneID: PaneID) {
-        record(scope: .pane(paneID)) { summary in
-            summary.storeFastPathCount += 1
-        }
-    }
-
-    func recordInvalidation(paneID: PaneID, impacts: WorklaneAuxiliaryInvalidation) {
-        guard !impacts.isEmpty else {
-            return
-        }
-
-        record(scope: .pane(paneID)) { summary in
-            for key in Self.invalidationKeys(for: impacts) {
-                Self.increment(key, in: &summary.auxiliaryInvalidationCounts)
-            }
-        }
-    }
-
-    func recordRender(_ kind: TerminalDiagnosticsRenderKind, activePaneID: PaneID?) {
-        let scope: Scope = activePaneID.map { .pane($0) } ?? .runtime
-        record(scope: scope) { summary in
-            Self.increment(kind.rawValue, in: &summary.renderCounts)
-        }
-    }
-
-    func recordRuntimeTopology(
-        liveRuntimeCount: Int,
-        visiblePaneCount: Int,
-        windowVisible: Bool,
-        windowKey: Bool
-    ) {
-        record(scope: .runtime) { summary in
-            summary.liveRuntimeCount = liveRuntimeCount
-            summary.visiblePaneCount = visiblePaneCount
-            summary.windowVisible = windowVisible
-            summary.windowKey = windowKey
-        }
-    }
-
-    private func record(scope: Scope, update: @escaping @Sendable (inout BurstSummary) -> Void) {
-        guard isEnabled else {
-            return
-        }
-
-        queue.async { [weak self] in
-            guard let self else {
-                return
-            }
-
-            let now = Date()
-            var burst = self.burstsByScope.removeValue(forKey: scope) ?? BurstState(
-                summary: BurstSummary(scope: scope, startedAt: now, endedAt: now),
-                lastEventAt: now,
-                flushWorkItem: nil
-            )
-
-            if now.timeIntervalSince(burst.lastEventAt) > self.quietPeriod {
-                burst.flushWorkItem?.cancel()
-                self.emit(burst.summary, endedAt: burst.lastEventAt)
-                burst = BurstState(
-                    summary: BurstSummary(scope: scope, startedAt: now, endedAt: now),
-                    lastEventAt: now,
-                    flushWorkItem: nil
-                )
-            }
-
-            update(&burst.summary)
-            burst.lastEventAt = now
-            burst.summary.endedAt = now
-            burst.flushWorkItem?.cancel()
-            let flushWorkItem = DispatchWorkItem { [weak self] in
-                self?.flush(scope: scope)
-            }
-            burst.flushWorkItem = flushWorkItem
-            self.burstsByScope[scope] = burst
-            self.queue.asyncAfter(deadline: .now() + self.quietPeriod, execute: flushWorkItem)
-        }
-    }
-
-    private func flush(scope: Scope) {
-        guard let burst = burstsByScope.removeValue(forKey: scope) else {
-            return
-        }
-
-        emit(burst.summary, endedAt: burst.lastEventAt)
-    }
-
-    private func emit(_ summary: BurstSummary, endedAt: Date) {
-        var summary = summary
-        summary.endedAt = endedAt
-        terminalDiagnosticsLogger.log(
-            "burst scope=\(summary.scope.label, privacy: .public) payload=\(Self.logPayload(for: summary), privacy: .public)"
-        )
-
-        let onEmit = self.onEmit
-        deliverEmission {
-            onEmit?(summary)
-        }
-    }
-
-    func flushForTesting() {
-        let pendingBursts: [(summary: BurstSummary, endedAt: Date)] = queue.sync {
-            let pendingBursts = burstsByScope.values.map { burst in
-                burst.flushWorkItem?.cancel()
-                return (summary: burst.summary, endedAt: burst.lastEventAt)
-            }
-            burstsByScope.removeAll()
-            return pendingBursts
-        }
-
-        pendingBursts.forEach { burst in
-            emit(burst.summary, endedAt: burst.endedAt)
-        }
-    }
-
-    private var isEnabled: Bool {
-        enabledLock.lock()
-        defer { enabledLock.unlock() }
-        return enabled
-    }
-
-    private static func milliseconds(from nanoseconds: UInt64) -> Double {
-        Double(nanoseconds) / 1_000_000
-    }
-
-    private static var defaultSharedEnabled: Bool {
-        #if DEBUG
-        ProcessInfo.processInfo.environment["ZENTTY_TERMINAL_DIAGNOSTICS"] == "1"
-        #else
-        false
-        #endif
-    }
-
-    private static func rowsBelowViewport(total: UInt64, offset: Double, visible: UInt64) -> Double {
-        let clampedOffset = min(max(offset, 0), max(0, Double(total) - Double(visible)))
-        let remainingRows = Double(total) - clampedOffset
-        let visibleRows = min(Double(visible), max(0, remainingRows))
-        return max(0, remainingRows - visibleRows)
-    }
-
-    private static func metadataChangeKindKey(_ kind: TerminalMetadataChangeKind) -> String {
-        switch kind {
-        case .noop:
-            return "noop"
-        case .volatileTitleOnly:
-            return "volatileTitleOnly"
-        case .meaningful:
-            return "meaningful"
-        }
-    }
-
-    private static func toolKey(_ tool: AgentTool?) -> String {
-        switch tool {
-        case .zentty:
-            return "zentty"
-        case .amp:
-            return "amp"
-        case .claudeCode:
-            return "claude"
-        case .codex:
-            return "codex"
-        case .copilot:
-            return "copilot"
-        case .cursor:
-            return "cursor"
-        case .droid:
-            return "droid"
-        case .gemini:
-            return "gemini"
-        case .kimi:
-            return "kimi"
-        case .openCode:
-            return "opencode"
-        case .pi:
-            return "pi"
-        case .omp:
-            return "omp"
-        case .grok:
-            return "grok"
-        case .agy:
-            return "agy"
-        case .hermes:
-            return "hermes"
-        case .vibe:
-            return "vibe"
-        case .smallHarness:
-            return "small-harness"
-        case .custom(let name):
-            return "custom:\(name)"
-        case nil:
-            return "none"
-        }
-    }
-
-    private static func actionPayloadKey(for payload: LibghosttySurfaceActionPayload) -> String {
-        switch payload {
-        case .setTitle:
-            return "title"
-        case .pwd:
-            return "pwd"
-        case .progressReport:
-            return "progressReport"
-        case .commandFinished,
-             .desktopNotification,
-             .startSearch,
-             .endSearch,
-             .searchTotal,
-             .searchSelected,
-             .openURL:
-            return "ordered"
-        case .scrollbar:
-            return "scrollbar"
-        case .cellSize:
-            return "cellSize"
-        case .mouseShape:
-            return "mouseShape"
-        }
-    }
-
-    private static func invalidationKeys(for impacts: WorklaneAuxiliaryInvalidation) -> [String] {
-        var keys: [String] = []
-        if impacts.contains(.sidebar) { keys.append("sidebar") }
-        if impacts.contains(.header) { keys.append("header") }
-        if impacts.contains(.canvas) { keys.append("canvas") }
-        if impacts.contains(.attention) { keys.append("attention") }
-        if impacts.contains(.openWith) { keys.append("openWith") }
-        if impacts.contains(.reviewRefresh) { keys.append("reviewRefresh") }
-        if impacts.contains(.surfaceActivities) { keys.append("surfaceActivities") }
-        return keys
-    }
-
-    private static func increment(_ key: String, in dictionary: inout [String: Int]) {
-        dictionary[key, default: 0] += 1
-    }
-
-    private static func logPayload(for summary: BurstSummary) -> String {
-        var components: [String] = [
-            "startedAt=\(summary.startedAt.timeIntervalSince1970)",
-            "endedAt=\(summary.endedAt.timeIntervalSince1970)",
-            "wakeupCount=\(summary.wakeupCount)",
-            "wakeupEnqueueCount=\(summary.wakeupEnqueueCount)",
-            "tickCount=\(summary.tickCount)",
-            "tickTotalMilliseconds=\(summary.tickTotalMilliseconds)",
-            "tickMaxMilliseconds=\(summary.tickMaxMilliseconds)",
-            "mainQueueDelayTotalMilliseconds=\(summary.mainQueueDelayTotalMilliseconds)",
-            "mainQueueDelayMaxMilliseconds=\(summary.mainQueueDelayMaxMilliseconds)",
-            "actionCallbackCount=\(summary.actionCallbackCount)",
-            "actionDrainCount=\(summary.actionDrainCount)",
-            "actionDrainQueueDelayTotalMilliseconds=\(summary.actionDrainQueueDelayTotalMilliseconds)",
-            "actionDrainQueueDelayMaxMilliseconds=\(summary.actionDrainQueueDelayMaxMilliseconds)",
-            "scrollbarApplyCount=\(summary.scrollbarApplyCount)",
-            "scrollbarApplyTotalMilliseconds=\(summary.scrollbarApplyTotalMilliseconds)",
-            "scrollbarApplyMaxMilliseconds=\(summary.scrollbarApplyMaxMilliseconds)",
-            "scrollHostSyncCount=\(summary.scrollHostSyncCount)",
-            "scrollHostSyncTotalMilliseconds=\(summary.scrollHostSyncTotalMilliseconds)",
-            "scrollHostSyncMaxMilliseconds=\(summary.scrollHostSyncMaxMilliseconds)",
-            "scrollbarGeometryApplyCount=\(summary.scrollbarGeometryApplyCount)",
-            "documentHeightChangeCount=\(summary.documentHeightChangeCount)",
-            "documentHeightMaxPoints=\(summary.documentHeightMaxPoints)",
-            "documentHeightMaxDeltaPoints=\(summary.documentHeightMaxDeltaPoints)",
-            "scrollbarBottomAlignedCount=\(summary.scrollbarBottomAlignedCount)",
-            "scrollbarOffBottomCount=\(summary.scrollbarOffBottomCount)",
-            "scrollbarWasAtBottomCount=\(summary.scrollbarWasAtBottomCount)",
-            "scrollbarAutoFollowEligibleCount=\(summary.scrollbarAutoFollowEligibleCount)",
-            "scrollbarAutoFollowSuppressedCount=\(summary.scrollbarAutoFollowSuppressedCount)",
-            "scrollbarAutoFollowAppliedCount=\(summary.scrollbarAutoFollowAppliedCount)",
-            "scrollbarUserScrolledAwayCount=\(summary.scrollbarUserScrolledAwayCount)",
-            "scrollbarExplicitSyncAllowedCount=\(summary.scrollbarExplicitSyncAllowedCount)",
-            "scrollbarMaxTotalRows=\(summary.scrollbarMaxTotalRows)",
-            "scrollbarMaxOffsetRows=\(summary.scrollbarMaxOffsetRows)",
-            "scrollbarMaxVisibleRows=\(summary.scrollbarMaxVisibleRows)",
-            "replaySuppressionEnterCount=\(summary.replaySuppressionEnterCount)",
-            "replaySuppressionExitCount=\(summary.replaySuppressionExitCount)",
-            "replaySuppressionTotalMilliseconds=\(summary.replaySuppressionTotalMilliseconds)",
-            "occlusionVisibleCount=\(summary.occlusionVisibleCount)",
-            "occlusionHiddenCount=\(summary.occlusionHiddenCount)",
-            "reflectScrolledClipViewCount=\(summary.reflectScrolledClipViewCount)",
-            "scrollToRowActionCount=\(summary.scrollToRowActionCount)",
-            "viewportSyncCount=\(summary.viewportSyncCount)",
-            "viewportSyncTotalMilliseconds=\(summary.viewportSyncTotalMilliseconds)",
-            "viewportSyncMaxMilliseconds=\(summary.viewportSyncMaxMilliseconds)",
-            "metadataDeliveryCount=\(summary.metadataDeliveryCount)",
-            "storeUpdateCount=\(summary.storeUpdateCount)",
-            "storeFastPathCount=\(summary.storeFastPathCount)",
-            "titleChangeCount=\(summary.titleChangeCount)",
-            "uniqueTitleCount=\(summary.uniqueTitleCount)"
-        ]
-
-        if let commandHint = summary.commandHint {
-            components.append("commandHint=\(commandHint)")
-        }
-        if let liveRuntimeCount = summary.liveRuntimeCount {
-            components.append("liveRuntimeCount=\(liveRuntimeCount)")
-        }
-        if let visiblePaneCount = summary.visiblePaneCount {
-            components.append("visiblePaneCount=\(visiblePaneCount)")
-        }
-        if let windowVisible = summary.windowVisible {
-            components.append("windowVisible=\(windowVisible)")
-        }
-        if let windowKey = summary.windowKey {
-            components.append("windowKey=\(windowKey)")
-        }
-        if let firstTitle = summary.firstTitle {
-            components.append("firstTitle=\(firstTitle)")
-        }
-        if let lastTitle = summary.lastTitle {
-            components.append("lastTitle=\(lastTitle)")
-        }
-        if let firstScrollbarPosition = summary.firstScrollbarPosition {
-            components.append("firstScrollbarPosition=\(firstScrollbarPosition)")
-        }
-        if let lastScrollbarPosition = summary.lastScrollbarPosition {
-            components.append("lastScrollbarPosition=\(lastScrollbarPosition)")
-        }
-        if let scrollbarMinOffsetRows = summary.scrollbarMinOffsetRows {
-            components.append("scrollbarMinOffsetRows=\(scrollbarMinOffsetRows)")
-        }
-        if let scrollbarMinVisibleRows = summary.scrollbarMinVisibleRows {
-            components.append("scrollbarMinVisibleRows=\(scrollbarMinVisibleRows)")
-        }
-        if !summary.replaySuppressionReasonCounts.isEmpty {
-            components.append("replaySuppressionReasons=\(summary.replaySuppressionReasonCounts)")
-        }
-        if !summary.metadataChangeKindCounts.isEmpty {
-            components.append("metadataChangeKinds=\(summary.metadataChangeKindCounts)")
-        }
-        if !summary.metadataToolCounts.isEmpty {
-            components.append("metadataTools=\(summary.metadataToolCounts)")
-        }
-        if !summary.actionPayloadCounts.isEmpty {
-            components.append("actionPayloads=\(summary.actionPayloadCounts)")
-        }
-        if !summary.auxiliaryInvalidationCounts.isEmpty {
-            components.append("invalidations=\(summary.auxiliaryInvalidationCounts)")
-        }
-        if !summary.renderCounts.isEmpty {
-            components.append("renders=\(summary.renderCounts)")
-        }
-
-        return components.joined(separator: " ")
-    }
-
-    #if DEBUG
-    static func logPayloadForTesting(_ summary: BurstSummary) -> String {
-        logPayload(for: summary)
-    }
-    #endif
 }
